# fix progress bar when step_size>1